Make xhr importable and use separately
Previous xhr fallback commit broke beacon function return promise type.
This commit reverts that and requires consumer to check Promise and
fetch existence and import xhr separately.
