+++ .vscode/settings.json
@@ -3,5 +3,8 @@
   "files.associations": {
     "api-extractor.json": "jsonc"
   },
-  "editor.formatOnSave": true
+  "editor.formatOnSave": true,
+  "cSpell.words": [
+    "Retryable"
+  ]
 }
