--- docs/beacon-transporter.beaconinit.beaconconfig.md
@@ -1,11 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [beacon-transporter](./beacon-transporter.md) &gt; [BeaconInit](./beacon-transporter.beaconinit.md) &gt; [beaconConfig](./beacon-transporter.beaconinit.beaconconfig.md)
-
-## BeaconInit.beaconConfig property
-
-<b>Signature:</b>
-
-```typescript
-beaconConfig?: BeaconConfig;
-```
