--- docs/beacon-transporter.beaconinit.compress.md
@@ -1,11 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [beacon-transporter](./beacon-transporter.md) &gt; [BeaconInit](./beacon-transporter.beaconinit.md) &gt; [compress](./beacon-transporter.beaconinit.compress.md)
-
-## BeaconInit.compress property
-
-<b>Signature:</b>
-
-```typescript
-compress?: boolean;
-```
