--- docs/beacon-transporter.beaconinitwithcustomdb.beaconconfig.md
@@ -1,11 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [beacon-transporter](./beacon-transporter.md) &gt; [BeaconInitWithCustomDB](./beacon-transporter.beaconinitwithcustomdb.md) &gt; [beaconConfig](./beacon-transporter.beaconinitwithcustomdb.beaconconfig.md)
-
-## BeaconInitWithCustomDB.beaconConfig property
-
-<b>Signature:</b>
-
-```typescript
-beaconConfig?: BeaconConfig;
-```
