--- docs/beacon-transporter.beaconinitwithcustomdb.compress.md
@@ -1,11 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [beacon-transporter](./beacon-transporter.md) &gt; [BeaconInitWithCustomDB](./beacon-transporter.beaconinitwithcustomdb.md) &gt; [compress](./beacon-transporter.beaconinitwithcustomdb.compress.md)
-
-## BeaconInitWithCustomDB.compress property
-
-<b>Signature:</b>
-
-```typescript
-compress?: boolean;
-```
