+++ docs/beacon-transporter.isglobalfetchsupported.md
@@ -1,0 +1,16 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [beacon-transporter](./beacon-transporter.md) &gt; [isGlobalFetchSupported](./beacon-transporter.isglobalfetchsupported.md)
+
+## isGlobalFetchSupported() function
+
+
+<b>Signature:</b>
+
+```typescript
+export declare function isGlobalFetchSupported(): boolean;
+```
+<b>Returns:</b>
+
+boolean
+
