+++ docs/beacon-transporter.iskeepalivefetchsupported.md
@@ -1,0 +1,16 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [beacon-transporter](./beacon-transporter.md) &gt; [isKeepaliveFetchSupported](./beacon-transporter.iskeepalivefetchsupported.md)
+
+## isKeepaliveFetchSupported() function
+
+
+<b>Signature:</b>
+
+```typescript
+export declare function isKeepaliveFetchSupported(): boolean;
+```
+<b>Returns:</b>
+
+boolean
+
