+++ docs/beacon-transporter.md
@@ -19,6 +19,9 @@ Transporting data to the server made easy. Support persistent retry and gzip com
 |  [createBeacon(init)](./beacon-transporter.createbeacon.md) |  |
 |  [createBeacon(init)](./beacon-transporter.createbeacon_1.md) |  |
 |  [createLocalStorageRetryDB({ keyName, maxNumber, throttleWait, headerName, attemptLimit, compressFetch, })](./beacon-transporter.createlocalstorageretrydb.md) |  |
+|  [isGlobalFetchSupported()](./beacon-transporter.isglobalfetchsupported.md) |  |
+|  [isKeepaliveFetchSupported()](./beacon-transporter.iskeepalivefetchsupported.md) |  |
+|  [xhr(url, body, options)](./beacon-transporter.xhr.md) |  |
 
 ## Interfaces
