+++ docs/beacon-transporter.xhr.md
@@ -1,0 +1,27 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [beacon-transporter](./beacon-transporter.md) &gt; [xhr](./beacon-transporter.xhr.md)
+
+## xhr() function
+
+
+<b>Signature:</b>
+
+```typescript
+export declare function xhr(url: string, body: string, options?: {
+    headers?: Record<string, string>;
+}): void;
+```
+
+## Parameters
+
+|  Parameter | Type | Description |
+|  --- | --- | --- |
+|  url | string |  |
+|  body | string |  |
+|  options | { headers?: Record&lt;string, string&gt;; } |  |
+
+<b>Returns:</b>
+
+void
+
