+++ src/index.ts
@@ -9,5 +9,6 @@ export {
   createLocalStorageRetryDB,
   LocalStorageRetryDB,
 } from './local-storage-retrydb';
+export * from './network';
 export { RetryDB } from './queue';
 export { gzipSync } from 'fflate';
